Move encode commands to encode.py
